Adds archive and timestamp_backups settings.
There is already a --no-archive command line option, but adding it to the settings means that it works without having to edit the provided backup_grafana.sh script.

Setting the timestamp_backups setting to false will mean that the output files do not include the timestamp in the name, but will instead delete and recreate existing files.  This is better if the backups are being kept in a version-control repository since you want to use the same filenames on each backup.

The restore function still needs to be modified to handle the output directory directly rather than a .tar.gz file.